Refactor attendance
Changes the timeout and intervals on data refresh to be related to
actions taken by the user. Specifically, by marking a camper as 'here'
or 'not here', the timer removes the old timer and starts a new one.
This helps avoid ghosting refreshes of data, where a camper could be
marked 'not here' before the refresh comes in, and be momentarily
displayed as 'here' before a new update happens.

This is so much cleaner now!!